add agregarPuntuacion operation

+++ src/Clases/Puntuacion/Puntuacion.cpp
@@ -8,15 +8,15 @@ using namespace std;
 class Puntuacion : public ICollectible
 {
 private:
-    int puntuacion;
+    double puntuacion;
     Videojuego *juego;
     Jugador *creador;
 
 public:
     Puntuacion(int, Videojuego *, Jugador *);
-    void setPuntuacion(int);
+    void setPuntuacion(double);
     void setCreador(Jugador *);
-    int getPuntuacion();
+    double getPuntuacion();
     Jugador *getCreador();
 };
 
@@ -27,7 +27,7 @@ Puntuacion::Puntuacion(int puntuacion, Videojuego *juego, Jugador *creador)
     this->creador = creador;
 }
 
-void Puntuacion::setPuntuacion(int puntuacion)
+void Puntuacion::setPuntuacion(double puntuacion)
 {
     this->puntuacion = puntuacion;
 }
@@ -37,7 +37,7 @@ void Puntuacion::setCreador(Jugador *creador)
     this->creador = creador;
 }
 
-int Puntuacion::getPuntuacion()
+double Puntuacion::getPuntuacion()
 {
     return this->puntuacion;
 }
